requirements.txt
pycryptodome~=3.23.0
cryptography~=46.0.5